test(offload): share component fixtures
